% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctNodeHeights.R
\name{correctNodeHeights}
\alias{correctNodeHeights}
\title{correctNodeHeights}
\usage{
correctNodeHeights(CAET, sd = 0.01)
}
\arguments{
\item{CAET}{List or singular cladogenetic events or anagenetic events tables from a BSM analysis.}

\item{sd}{Standard deiation of normal deviate to add to node heights}
}
\value{
CAET
}
\description{
adds a small random number to each node height so that none are matching
}
\author{
Alex Skeels
}
