% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventTiming.R
\name{getEventTiming}
\alias{getEventTiming}
\title{getEventsTiming}
\usage{
getEventTiming(tree, CET, AET)
}
\arguments{
\item{tree}{The phylogeny used in BSM analysis.}

\item{CET}{The cladogenetic events table output from BSM analysis.}

\item{AET}{The anagenetic events table output from BSM analysis.}
}
\value{
Returns the events timing table as a data frame.
}
\description{
Extracts the events timing information from the anagenetic and Cladogenetic events tables from a BGB analysis and stores it in the events timing table.
The events timing table contains the time of each event,
the node at which the event is located on the phylogeny (or beneath which for an anagenetic event along a branch),
whether the event was cladogenetic or anagenetic, the number of species extant at that time,
and a running total of the number of species present in each unique state combination.
}
\author{
Alex Skeels
}
