% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLTSTDataTable.R
\name{getLTSTDataTable}
\alias{getLTSTDataTable}
\title{getLTSTDataTable}
\usage{
getLTSTDataTable(ETT, rangeStates)
}
\arguments{
\item{ETT}{The events timing table (from getEventsTiming).}

\item{rangeStates}{The lookup table for geographic range states (from getRangeState).}
}
\value{
An LTST data table as class 'data frame'
}
\description{
Creates a lineages through space and time data table by amalgamating the running diversity count across state combination identifiers into their component geographic areas.
}
\author{
Alex Skeels
}
