% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLTSTFromStochasticMaps.R
\name{getLTSTFromStochasticMaps}
\alias{getLTSTFromStochasticMaps}
\title{getLTSTFromStochasticMaps}
\usage{
getLTSTFromStochasticMaps(tr, CET_list, AET_list, tipranges)
}
\arguments{
\item{tr}{Phylogenetic tree used in BGB analysis.}

\item{CET_list}{List of cladogenetic events tables from a BSM analysis.}

\item{AET_list}{List of Anagenetic events tables from a BSM analysis.}
}
\value{
A list of LTST data tables.
}
\description{
Applies getLTSTDataTable across two lists of cladogenetic and anagenetic events tables from a biegoegraphic stochastic mapping analysis (BSM).
This function iterates over each stochastic map in the list, producing a list of LTST data tables.
}
\author{
Alex Skeels
}
