% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRangeState.R
\name{getRangeState}
\alias{getRangeState}
\title{getRangeState}
\usage{
getRangeState(CET, AET, tipranges)
}
\arguments{
\item{CET}{The cladogenetic events table output from BSM analysis.}

\item{AET}{The anagenetic events table output from BSM analysis.}

\item{tipranges}{The tipranges object (class 'tipranges') from BGB analysis containing a presence absence matrix with species names as rows and geographic states in columns.}
}
\value{
A lookup table as class 'data frame'.
}
\description{
Generates lookup table to match geographic range states (as charater strings) to state combination identifiers (as numeric).
}
\author{
Alex Skeels
}
