% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventTiming.R
\name{getTableIndexFromEventTime}
\alias{getTableIndexFromEventTime}
\title{getTableIndexFromEventTime}
\usage{
getTableIndexFromEventTime(CET, AET, ET, event.type, ETT)
}
\arguments{
\item{CET}{the cladogenetic events table output from BGB analysis}

\item{AET}{the anagenetic events table output from BGB analysis}

\item{event.type}{clado or ana}

\item{ETT}{events timing table}
}
\value{
logical (TRUE for cladogenetic or FALSE for anagenetic)
}
\description{
for either the CET or AET finds the index (row number) and node number for the current event using the event time as a reference
}
\author{
Alex Skeels
}
