% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventTiming.R
\name{isEventClado}
\alias{isEventClado}
\title{isEventClado}
\usage{
isEventClado(event.time, CET, AET)
}
\arguments{
\item{event.time}{timing of the current event}

\item{CET}{the cladogenetic events table output from BGB analysis}

\item{AET}{the anagenetic events table output from BGB analysis}
}
\value{
logical (TRUE for cladogenetic or FALSE for anagenetic)
}
\description{
asks whether the current event is cladogenetic or anagenetic
}
\author{
Alex Skeels
}
