% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventTiming.R
\name{nextEventTime}
\alias{nextEventTime}
\title{nextEventTime}
\usage{
nextEventTime(CET, AET, ET)
}
\arguments{
\item{CET}{cladogenetic event table}

\item{AET}{anagenetic event table}

\item{ET}{current event time}
}
\value{
returns the next event time in chronological order
}
\description{
this function takes the current event time and finds which event time folows it chronological by indexing the AET and CET
}
\author{
Alex Skeels
}
