% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventTiming.R
\name{setUpETT}
\alias{setUpETT}
\title{setUpETT}
\usage{
setUpETT(states, CET)
}
\arguments{
\item{states}{vector of named character states}

\item{CET}{the cladogenetic events table output from BGB analysis}
}
\value{
an empty ETT data frame
}
\description{
creates the initial (unfilled) events timing table from a vector of character states
}
\author{
Alex Skeels
}
