% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingWindowForStochasticMaps.R
\name{slidingWindowForStochasticMaps}
\alias{slidingWindowForStochasticMaps}
\title{slidingWindowForStochasticMaps}
\usage{
slidingWindowForStochasticMaps(LTST.dt, window = 0.1)
}
\arguments{
\item{LTST.dt}{Lineages through space and time data table.}

\item{window}{Sets the window size. default = 0.1.}

\item{tr}{Phylogenetic tree used in BSM analysis.}
}
\value{
Returns an LTST data table with standardised event times.
}
\description{
This function performs a sliding window analysis to obtain the number of species present in each region within regular time intervals based on the window size. This allows for comparison across biogeographic stochastic maps.
The default window size is 0.1 time units, but users should select values that make sense for their particular datasets.
}
\author{
Alex Skeels
}
