% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeRangeAcrossStochasticMaps.R
\name{timeRangeAcrossStochasticMaps}
\alias{timeRangeAcrossStochasticMaps}
\title{timeRangeAcrossStochasticMaps}
\usage{
timeRangeAcrossStochasticMaps(tr, window.LTST, melt = T)
}
\arguments{
\item{tr}{Phylogenetic tree used in BSM analysis.}

\item{window.LTST}{Lineages through space and time data table with standardised event times (from slidingWindowForStochasticMaps).}

\item{melt}{Whether to melt the results (useful for plotting with ggplot2)}
}
\value{
Diversity data for each geographic region summarised as quantiles in a data frame.
}
\description{
Summarises the upper (97.5%) and lower (2.5%) quantiles of species diversity count at each time interval across biogeographic stochastic maps (BSMs).
The summarised data can be returned in 'melted' data frame with five columns: time, region, lowerQ, uperrQ, mid.
Alternatively the summarised data can be returned in a data frame with a column for time, and a further two columns for each geographic region summarising the upper and lower quantile of diversity at each time point.
}
\author{
Alex Skeels
}
