% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventTiming.R
\name{whichStateChangesAna}
\alias{whichStateChangesAna}
\title{whichStateChangesAna}
\usage{
whichStateChangesAna(x, ETT, ET, index)
}
\arguments{
\item{x}{the AET}

\item{ETT}{events timing table}

\item{ET}{current event time}

\item{index}{index of AET}
}
\value{
updated ETT with changes made to the geographic states based on which event is occuring
}
\description{
modifies ETT by accounting for state changes along branches in the phylogeny (anagenetic changes).
The values of the originally occupied state will decrease by 1 and values of the new state will increase by 1.
}
\author{
Alex Skeels
}
