% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEventTiming.R
\name{whichStateChangesClado}
\alias{whichStateChangesClado}
\title{whichStateChangesClado}
\usage{
whichStateChangesClado(x, ETT, ET, index)
}
\arguments{
\item{x}{the CET}

\item{ETT}{events timing table}

\item{ET}{current event time}

\item{index}{index of CET}
}
\value{
updated ETT with changes made to the geographic states based on which event is occuring
}
\description{
amodifies ETT by accounting for state changes occur at nodes in the phylogeny - cladogenetic events.
A state can increase in diversity by 1 if new daughter species also occurs in the same state.
A different state from the parent can increase by 1 if new daughter species occupies a new state.
A state can decrease by 1 if both daughter species occupy different states from the parent - in which case those new states will each +1
}
\author{
Alex Skeels
}
